requirements.txt
python-telegram-bot>=21.0
PyYAML>=6.0
requests>=2.28